check_twemproxy only returns errors if requests have been attempted
The check would return errors for shards with no active
connections. This is a valid state for a shard to be in,
as Twemproxy will disconnect shards that have no activity.

Instead, we check twice and only report errors if there have
been requests to a shard that is down.

Also, the check has been translated entirely into ruby.